extract linux distribution file
