+++ conoha_client/features/image/domain/__init__.py
@@ -1,2 +1,4 @@
 """domain init."""
+from .distribution import Application, Distribution, DistVersion  # noqa: F401
 from .image import Image, ImageType  # noqa: F401
+from .operating_system import FileSystem  # noqa: F401
